+++ src/main/kotlin/co/yappuworld/board/client/application/dto/response/NoticeOverviewAppResponseDto.kt
@@ -1,4 +1,4 @@
-package co.yappuworld.board.application.dto.response
+package co.yappuworld.board.client.application.dto.response
 
 import co.yappuworld.board.domain.model.Notice
 import co.yappuworld.board.domain.vo.NoticeType
@@ -7,42 +7,43 @@ import co.yappuworld.user.infrastructure.model.UserWithLastActivityUnit
 import java.time.LocalDate
 import java.util.UUID
 
-data class NoticeAppResponseDto(
-    val notice: NoticeDetailsAppResponseDto,
-    val writer: NoticeDetailsWriterAppResponseDto
+data class NoticeOverviewAppResponseDto(
+    val notice: NoticeSimpleAppResponseDto,
+    val writer: NoticeOverviewWriterAppResponseDto
 ) {
+
     constructor(notice: Notice, user: UserWithLastActivityUnit) : this(
-        notice = NoticeDetailsAppResponseDto(notice),
-        writer = NoticeDetailsWriterAppResponseDto(user)
+        notice = NoticeSimpleAppResponseDto(notice),
+        writer = NoticeOverviewWriterAppResponseDto(user)
     )
 }
 
-data class NoticeDetailsAppResponseDto(
+data class NoticeSimpleAppResponseDto(
     val id: UUID,
+    val createdAt: LocalDate,
     val title: String,
     val content: String,
-    val createdAt: LocalDate,
-    val type: NoticeType
+    val noticeType: NoticeType
 ) {
 
     constructor(notice: Notice) : this(
         id = notice.id,
+        createdAt = LocalDate.now(),
         title = notice.title,
-        content = notice.content,
-        createdAt = notice.createdAt.toLocalDate(),
-        type = notice.noticeType
+        content = notice.contentSummary.take(200),
+        noticeType = notice.noticeType
     )
 }
 
-data class NoticeDetailsWriterAppResponseDto(
-    val id: UUID,
+data class NoticeOverviewWriterAppResponseDto(
+    val userId: UUID,
     val name: String,
     val activityUnitGeneration: Int,
     val activityUnitPosition: Position
 ) {
 
     constructor(user: UserWithLastActivityUnit) : this(
-        id = user.userId,
+        userId = user.userId,
         name = user.name,
         activityUnitGeneration = user.activityUnit.generation,
         activityUnitPosition = user.activityUnit.position
